scoring-service/pyproject.toml
[project]
name = "scoring-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Zander1264",email = "87423331+Zander1264@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "ruff (>=0.13.3,<0.14.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)",
    "aiokafka (>=0.12.0,<0.13.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)",
    "opentelemetry-api (>=1.38.0,<2.0.0)",
    "opentelemetry-exporter-otlp (>=1.38.0,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.59b0,<0.60)",
    "opentelemetry-instrumentation-httpx (>=0.59b0,<0.60)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version = "py312"
line-length = 88
exclude = [
    ".eggs",
    ".git",
    ".mypy_cache",
    ".pytype",
    ".ruff_cache",
    ".venv",
    "__pypackages__",
    "__pycache__",
    "_build",
    "build",
    "dist",
    "venv",
    "tests",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "RET", # flake8-return
    "S",   # flake8-bandit
    "N",   # pep8-naming
    "DTZ", # flake8-datetimez
    "ERA", # eradicate
    "PD",  # pandas-vet
    "PGH", # pygrep-hooks
    "FLY", # flynt
    "PIE", # flake8-pie
    "T20", # flake8-print
    "PT",  # flake8-pytest-style
    "Q",   # flake8-quotes
    "RSE", # flake8-raise
    "RUF", # Ruff-specific rules
]
ignore = [
    "B008",   # do not perform function calls in argument defaults
    "S104",   # possible binding to all interfaces (for development server)
    "RUF003",
    "RUF002"
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"
multiline-quotes = "double"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
explicit_package_bases = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "src.tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false